Implemented, and recreated, Gun Mechanics
Shoots projectiles on left click. Ammo type can be switched with 1, 2 and 3.

Projectiles bounce on most objects but get destroyed when hitting animals or the ground.

Right click when held activates a short range vacuum that sucks suckable harvest objects into the gun, refilling the ammo.

Ammo accurately updates on the HUD.