0.7.4 Changes
Added vacuum ability to the gun.
Added harvest and regrowing system so that ammo can be restored to the gun through the farm.
Added the ability for animals to drop produce when they're happy so that the player can naturally obtain produce to sell through intended gameplay mechanics.
I finished Harrison's frantic animal implementation, adding the ability for animals to hit the player, knocking all the produce out of them as well as being able to pacify them and pick them up to return them to the pen.
Also implemented the tutorial level with bespoke scripts for it as well as scripts modified to suit the specific sections of the tutorial.